Rename gCPPProcess.h as CPPProcess.h. Implement general strategy: put code in .cc files, use .cu files as symlinks (and correspondingly, use the cc names of headers for .h files).
